Source code for Alertgy demo app
This app is not meant to serve as a basis for the final production app.
It’s little more than a carefully programmed slideshow designed to
showcase the final app’s planned ability to graph glucose data, trigger
an alarm in response to low glucose, and send a text message alert to
an emergency contact if that alert does not receive a response from the
user.